+++ backend/src/middlewares/utility.ts
@@ -1,7 +1,8 @@
 import _ from "lodash";
-import type { Response, NextFunction, RequestHandler } from "express";
+import type { Request, Response, NextFunction, RequestHandler } from "express";
 import { handleMonkeyResponse, MonkeyResponse } from "../utils/monkey-response";
 import { isDevEnvironment } from "../utils/misc";
+import { recordClientVersion as prometheusRecordClientVersion } from "../utils/prometheus";
 
 export const emptyMiddleware = (
   _req: MonkeyTypes.Request,
@@ -44,4 +45,19 @@ export function useInProduction(
   return middlewares.map((middleware) =>
     isDevEnvironment() ? emptyMiddleware : middleware
   );
+}
+
+/**
+ * record the client version from the `x-client-version`  or ` client-version` header to prometheus
+ */
+export function recordClientVersion(): RequestHandler {
+  return (req: Request, _res: Response, next: NextFunction) => {
+    const clientVersion =
+      (req.headers["x-client-version"] as string) ||
+      req.headers["client-version"];
+
+    prometheusRecordClientVersion(clientVersion?.toString() ?? "unknown");
+
+    next();
+  };
 }
